refactor: make vote views extend same base class
