+++ apps/votes_results/views/__init__.py
@@ -1,8 +1,16 @@
+# all votable polls view
+from .votable_polls_view import *
+
+# vote views
 from .vote.single_option_vote_view import *
 from .vote.majority_judgment_vote_view import *
-from .votable_polls_view import *
-from .generic_vote_view import *
-from .results.single_option_results_view import *
-from .results.majority_judgment_results_view import *
+from .vote.generic_vote_view import *
+
+# vote recap views
 from .vote_recap.single_option_recap_view import *
 from .vote_recap.majority_judgment_recap_view import *
+
+# results views
+from .results.single_option_results_view import *
+from .results.majority_judgment_results_view import *
+from .results.generic_results_view import *
