+++ apps/votes_results/views/vote/generic_vote_view.py
@@ -5,12 +5,7 @@ from apps.polls_management.classes.poll_token_validation.token_validation import
 from apps.polls_management.models.poll_model import PollModel
 from apps.polls_management.services.poll_service import PollService
 from apps.polls_management.services.poll_token_service import PollTokenService
-from apps.polls_management.models.poll_token import PollTokens
-from sesame.decorators import authenticate
-from sesame.utils import get_user, get_token
-from django.contrib.auth import logout
 
-# @authenticate(required=False, scope="Poll:{poll_id}")
 def generic_vote_view(request, poll_id: int):
     """Redirect to poll's main vote method"""
 
@@ -89,22 +84,3 @@ def generic_vote_view(request, poll_id: int):
             return HttpResponseRedirect(reverse(
                 'apps.votes_results:single_option_vote', 
                 args=(poll_id,)))
-
-
-def generic_results_view(request, poll_id: int):
-    """Redirect to poll's main vote method's results"""
-    
-    try:
-        # Retrieve poll
-        poll: PollModel = PollService.get_poll_by_id(poll_id)
-    except Exception:
-        raise Http404(f"Poll with id {poll_id} not found.")
-    
-    # redirect to proper vote method's results
-    if poll.poll_type == PollModel.PollType.MAJORITY_JUDJMENT:
-        return HttpResponseRedirect(reverse('apps.votes_results:majority_judgment_results', args=(poll_id,)))
-    else:
-        return HttpResponseRedirect(reverse('apps.votes_results:single_option_results', args=(poll_id,)))
-
-
-# TODO: (maybe) use those pages to handle generalized parts (?)
