+++ apps/votes_results/views/vote/vote_view_schema.py
@@ -1,0 +1,18 @@
+import abc
+
+from django.views import View
+
+
+class VoteViewSchema(abc.ABC, View):
+    """
+    Abstract schema for a vote view. 
+
+    It implements (on high level) the management of get and post requests, 
+    using ... + some template methods.
+    """
+
+    def get(self, request, *args, **kwargs):
+        pass
+
+    def post(self, request, *args, **kwargs):
+        pass
